ConstitutionalAI/requirements.txt
fastapi
uvicorn
langchain
langchain_groq
langchain_google_genai
langchain_community
python-dotenv
faiss-cpu
langchain-huggingface
sentencepiece
transformers 
huggingface_hub
pypdf
duckduckgo_search
httpx
pandas
newsapi-python
requests
unstructured
langchain-google-genai
markdown-it-py